Move trimerge-sync-basic-client, trimerge-sync-basic-server and example-broadcast-channel to a new 'example' directory.
