+++ example/example-broadcast-channel/src/AppDoc.ts
