+++ example/example-broadcast-channel/src/components/ClientColor.ts
