+++ example/example-broadcast-channel/src/lib/trimergeNumber.ts
