+++ example/trimerge-sync-basic-client/package.json
