+++ example/trimerge-sync-basic-client/rollup.config.js
