+++ example/trimerge-sync-basic-client/src/index.ts
