+++ example/trimerge-sync-basic-client/tsconfig.json
