+++ example/trimerge-sync-basic-server/src/lib/docs.ts
