+++ example/trimerge-sync-basic-server/src/types.ts
