+++ example/trimerge-sync-basic-server/tsconfig.dev-start.json
