+++ example/trimerge-sync-basic-server/tsconfig.json
