+++ package.json
@@ -1,7 +1,8 @@
 {
   "private": true,
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "example/*"
   ],
   "scripts": {
     "test-ci": "CI=true wsrun --no-prefix --serial test-ci",
